refactor(signals): move signal handlers to platform-specific files
